Created encryptstr
Notes:
This takes in a string
This uses AES.CFB (Cipher Feedback). It uses a random value called the IV plus the key for encryption. IV = nonce, 16 random bytes generated by the OS. The nonce is publicly shared and is needed to decrypt.
This also verifies the string with MD5, which creates a hash of the plaintext. It returns this a a tag.
This returns the ciphertext, the tag, and the nonce as a tuple